Create new utils files
