+++ DevToolkt/reactiveDom/src/jsMain/kotlin/dev/toolkt/dom/reactive/utils/html/HTMLElementUtils.kt
@@ -2,17 +2,11 @@ package dev.toolkt.dom.reactive.utils.html
 
 import dev.toolkt.dom.reactive.style.ReactiveStyle
 import dev.toolkt.dom.reactive.utils.createReactiveElement
-import dev.toolkt.dom.reactive.utils.event.offsetPoint
-import dev.toolkt.dom.reactive.widget.Button
-import dev.toolkt.geometry.Point
-import dev.toolkt.reactive.cell.Cell
 import dev.toolkt.reactive.event_stream.EventStream
 import dev.toolkt.reactive.event_stream.cast
 import dev.toolkt.reactive.event_stream.getEventStream
-import dev.toolkt.reactive.event_stream.hold
 import dev.toolkt.reactive.reactive_list.ReactiveList
 import org.w3c.dom.Document
-import org.w3c.dom.Element
 import org.w3c.dom.HTMLButtonElement
 import org.w3c.dom.HTMLDivElement
 import org.w3c.dom.HTMLElement
@@ -46,36 +40,6 @@ fun Document.createReactiveHtmlSpanElement(
     style = style,
     children = children,
 ) as HTMLSpanElement
-
-fun Element.getMouseEnterEventStream(): EventStream<MouseEvent> = this.getEventStream(
-    type = "mouseenter",
-).cast()
-
-fun Element.getMouseDownEventStream(): EventStream<MouseEvent> = this.getEventStream(
-    type = "mousedown",
-).cast()
-
-fun Element.getMouseDownEventStream(
-    button: Short,
-): EventStream<MouseEvent> = this.getMouseDownEventStream().filter { it.button == button }
-
-fun Element.getMouseUpEventStream(): EventStream<MouseEvent> = this.getEventStream(
-    type = "mouseup",
-).cast()
-
-fun Element.getMouseUpEventStream(
-    button: Short,
-): EventStream<MouseEvent> = this.getMouseUpEventStream().filter { it.button == button }
-
-fun Element.getMouseMoveEventStream(): EventStream<MouseEvent> = this.getEventStream(
-    type = "mousemove",
-).cast()
-
-fun Element.getMouseLeaveEventStream(): EventStream<MouseEvent> = this.getEventStream(
-    type = "mouseleave",
-).cast()
-
-fun Element.getMouseOffsetCell(): Cell<Point?> = getMouseMoveEventStream().map { it.offsetPoint }.hold(null)
 
 fun HTMLElement.getClickEventStream(): EventStream<MouseEvent> = this.getEventStream(
     type = "click"
